Update to move php fpm word press integration tests out
