+++ tests/Oryx.Integration.Tests/Php/PhpWordPressTest.cs
@@ -25,59 +25,6 @@ namespace Microsoft.Oryx.Integration.Tests
         }
 
         [Theory]
-        [InlineData("8.0-fpm")]
-        [InlineData("7.4-fpm")]
-        [InlineData("7.3-fpm")]
-        public async Task PhpFpmWithWordPress56(string phpVersion)
-        {
-            // Arrange
-            string hostDir = Path.Combine(_tempRootDir, Guid.NewGuid().ToString("N"));
-            var phpimageVersion = phpVersion.Split("-");
-
-            if (!Directory.Exists(hostDir))
-            {
-                Directory.CreateDirectory(hostDir);
-                using (var webClient = new WebClient())
-                {
-                    var wpZipPath = Path.Combine(hostDir, "wp.zip");
-                    webClient.DownloadFile("https://wordpress.org/wordpress-5.6.zip", wpZipPath);
-                    // The ZIP already contains a `wordpress` folder
-                    ZipFile.ExtractToDirectory(wpZipPath, hostDir);
-                }
-            }
-
-            var appName = "wordpress";
-            var volume = DockerVolume.CreateMirror(Path.Combine(hostDir, "wordpress"));
-            var appDir = volume.ContainerDir;
-            var appOutputDirVolume = CreateAppOutputDirVolume();
-            var appOutputDir = appOutputDirVolume.ContainerDir;
-            var buildScript = new ShellScriptBuilder()
-               .AddCommand($"oryx build {appDir} -i /tmp/int -o {appOutputDir} " +
-               $"--platform {PhpConstants.PlatformName} --platform-version {phpimageVersion[0]}")
-               .ToString();
-            var runScript = new ShellScriptBuilder()
-                .AddCommand("mkdir -p /home/site/wwwroot/")
-                .AddCommand($"cp -rf {appOutputDir}/* /home/site/wwwroot/")
-                .AddCommand("ls -la /home/site/wwwroot/")
-                .AddCommand($"oryx create-script -appPath /home/site/wwwroot/ -bindPort {ContainerPort} -output {RunScriptPath}")
-                .AddCommand(RunScriptPath)
-                .ToString();
-
-            // Act & Assert
-            await EndToEndTestHelper.BuildRunAndAssertAppAsync(
-                appName, _output, new[] { volume, appOutputDirVolume },
-                "/bin/sh", new[] { "-c", buildScript },
-                _imageHelper.GetRuntimeImage("php", phpVersion),
-                ContainerPort,
-                "/bin/sh", new[] { "-c", runScript },
-                async (hostPort) =>
-                {
-                    var data = await _httpClient.GetStringAsync($"http://localhost:{hostPort}/");
-                    Assert.Contains("<title>WordPress &rsaquo; Setup Configuration File</title>", data);
-                });
-        }
-
-        [Theory]
         [InlineData("7.3")]
         [InlineData("7.2")]
         [InlineData("7.0")]
